+++ Kernel/cpan-lib/XML/Parser/Lite.pm
@@ -5,8 +5,6 @@
 # SOAP::Lite is free software; you can redistribute it
 # and/or modify it under the same terms as Perl itself.
 #
-# $Id$
-#
 # ======================================================================
 
 package XML::Parser::Lite;
@@ -14,7 +12,7 @@ package XML::Parser::Lite;
 use strict;
 use warnings;
 
-our $VERSION = 0.717;
+our $VERSION = 0.719;
 
 sub new {
     my $class = shift;
@@ -249,7 +247,7 @@ XML::Parser::Lite - Lightweight regexp-based XML parser
 
 This Perl implements an XML parser with a interface similar to
 XML::Parser. Though not all callbacks are supported, you should be able to
-use it in the same way you use XML::Parser. Due to using experimantal regexp
+use it in the same way you use XML::Parser. Due to using experimental regexp
 features it'll work only on Perl 5.6 and above and may behave differently on
 different platforms.
 
@@ -264,7 +262,7 @@ not re-entrant).
 Constructor.
 
 As (almost) all SOAP::Lite constructors, new() returns the object called on
-when called as object method. This means that the following effectifely is
+when called as object method. This means that the following effectively is
 a no-op if $obj is a object:
 
  $obj = $obj->new();